Completed Readme.  Restructured Files.
