+++ CodeChallenges/CodeChallengeTests/CodeChallengeTests.csproj
@@ -23,6 +23,7 @@
     <ProjectReference Include="..\..\DataStructures\Implementations\Implementations.csproj" />
     <ProjectReference Include="..\LinkedLists\LinkedLists.csproj" />
     <ProjectReference Include="..\QueueWithStacks2\QueueWithStacks.csproj" />
+    <ProjectReference Include="..\QueueWithStacks\QueueWithStacks.csproj" />
   </ItemGroup>
 
 </Project>
