+++ CodeChallenges/QueueWithStacks/QueueWithStacks.csproj
@@ -1,12 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\DataStructures\Implementations\Implementations.csproj" />
-    <ProjectReference Include="..\QueueWithStacks2\QueueWithStacks2.csproj" />
   </ItemGroup>
 
 </Project>
