--- CodeChallenges/QueueWithStacks2/QueueWithStacks.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\DataStructures\Implementations\Implementations.csproj" />
-  </ItemGroup>
-
-</Project>
