+++ dotnet.sln
@@ -20,9 +20,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueueWithStacks", "CodeChallenges\QueueWithStacks2\QueueWithStacks.csproj", "{D6DFDCC4-B9A1-4885-84E4-35521028AF0D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExecutionTesting", "CodeChallenges\ExecutionTesting\ExecutionTesting.csproj", "{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QueueWithStacks", "CodeChallenges\QueueWithStacks\QueueWithStacks.csproj", "{CDA4EC1A-EC7B-49AD-8812-DCED18559258}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,14 +44,10 @@ Global
 		{E74AD3A8-A30F-4B81-A504-5B0902435AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E74AD3A8-A30F-4B81-A504-5B0902435AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E74AD3A8-A30F-4B81-A504-5B0902435AA0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D6DFDCC4-B9A1-4885-84E4-35521028AF0D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D6DFDCC4-B9A1-4885-84E4-35521028AF0D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D6DFDCC4-B9A1-4885-84E4-35521028AF0D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D6DFDCC4-B9A1-4885-84E4-35521028AF0D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CDA4EC1A-EC7B-49AD-8812-DCED18559258}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CDA4EC1A-EC7B-49AD-8812-DCED18559258}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CDA4EC1A-EC7B-49AD-8812-DCED18559258}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CDA4EC1A-EC7B-49AD-8812-DCED18559258}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -63,8 +57,7 @@ Global
 		{D87BE769-ACC8-46C4-BBA7-7F0B0159D356} = {08F49070-8D0C-4696-8E3E-FDBAA6A41FAB}
 		{72FF6C82-946F-498A-8AD7-C1E6ABA0EA9C} = {1BB78EB1-B9A8-4E82-BFCD-CADF0140E745}
 		{E74AD3A8-A30F-4B81-A504-5B0902435AA0} = {1BB78EB1-B9A8-4E82-BFCD-CADF0140E745}
-		{D6DFDCC4-B9A1-4885-84E4-35521028AF0D} = {1BB78EB1-B9A8-4E82-BFCD-CADF0140E745}
-		{DBAD9DC2-E87B-4BD5-98A4-5E15B924D390} = {1BB78EB1-B9A8-4E82-BFCD-CADF0140E745}
+		{CDA4EC1A-EC7B-49AD-8812-DCED18559258} = {1BB78EB1-B9A8-4E82-BFCD-CADF0140E745}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8D8CBDA3-1372-4033-8C11-81DBE1CC27E8}
